Cortx_multiple_data_pod_shutdown_opt_3 (#1760)
* CORTX:33252: Update of multiple pod shutdown tests test_multi_data_pod_failure.py


* CORTX:33252: INitial commit


* CORTX:33252: INitial commit


* CORTX:33252: Codacy fix and 1st level review comment update


* CORTX:33252:  review comment update in Sonal's and Rahul's Comments


* CORTX:33252: MInor fix


* CORTX:33252: MInor fix


* CORTX:33252: MInor fix in copy download dict use


* CORTX:33252: MInor fix in copy download dict use
